+++ pkg/kong/consumer.go
@@ -2,7 +2,7 @@ package kong
 
 import (
 	v1 "github.com/apache/apisix-ingress-controller/pkg/types/apisix/v1"
-	"github.com/api7/kongtoapisix/pkg/utils"
+	"github.com/api7/kong-to-apisix/pkg/utils"
 )
 
 func MigrateConsumer(kongConfig *KongConfig, configYamlAll *[]utils.YamlItem) (*[]v1.Consumer, error) {
