+++ pkg/kong/plugin.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	v1 "github.com/apache/apisix-ingress-controller/pkg/types/apisix/v1"
-	"github.com/api7/kongtoapisix/pkg/utils"
+	"github.com/api7/kong-to-apisix/pkg/utils"
 )
 
 var pluginMap = map[string]func(p Plugin) (v1.Plugins, []utils.YamlItem, error){
