+++ pkg/kong/route.go
@@ -5,7 +5,7 @@ import (
 	"strconv"
 
 	v1 "github.com/apache/apisix-ingress-controller/pkg/types/apisix/v1"
-	"github.com/api7/kongtoapisix/pkg/utils"
+	"github.com/api7/kong-to-apisix/pkg/utils"
 )
 
 func MigrateRoute(kongConfig *KongConfig, configYamlAll *[]utils.YamlItem) (*[]v1.Route, error) {
