+++ test/e2e/e2e_suite_test.go
@@ -3,10 +3,12 @@ package e2e
 import (
 	"testing"
 
-	"github.com/api7/kongtoapisix/pkg/apisix"
-	"github.com/api7/kongtoapisix/pkg/utils"
+	"github.com/api7/kong-to-apisix/pkg/apisix"
+	"github.com/api7/kong-to-apisix/pkg/utils"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
+
+	e2eutils "github.com/api7/kong-to-apisix/test/e2e/utils"
 )
 
 func TestMigrate(t *testing.T) {
@@ -14,7 +16,7 @@ func TestMigrate(t *testing.T) {
 	if err := apisix.EnableAPISIXStandalone(&apisixConfig); err != nil {
 		panic(err)
 	}
-	if err := utils.AppendToConfigYaml(&apisixConfig, apisixConfigYamlPath); err != nil {
+	if err := utils.AppendToConfigYaml(&apisixConfig, e2eutils.ApisixConfigYamlPath); err != nil {
 		panic(err)
 	}
 	RegisterFailHandler(Fail)
