+++ test/e2e/upsteam_test.go
@@ -3,6 +3,7 @@ package e2e
 import (
 	"strings"
 
+	"github.com/api7/kong-to-apisix/test/e2e/utils"
 	"github.com/globocom/gokong"
 	"github.com/onsi/ginkgo"
 	"github.com/onsi/gomega"
@@ -13,44 +14,44 @@ var _ = ginkgo.Describe("upstream", func() {
 
 	ginkgo.JustBeforeEach(func() {
 		kongCli = gokong.NewClient(gokong.NewDefaultConfig())
-		err := purgeAll(kongCli)
+		err := utils.PurgeAll(kongCli)
 		gomega.Expect(err).To(gomega.BeNil())
 	})
 
 	ginkgo.It("default route, service and upstream", func() {
-		createdUpstream := defaultUpstream()
+		createdUpstream := utils.DefaultUpstream()
 		createdUpstream.Name = "upstream"
 		kongUpstream, err := kongCli.Upstreams().Create(createdUpstream)
 		gomega.Expect(err).To(gomega.BeNil())
 
-		createdTarget := defaultTarget()
-		createdTarget.Target = strings.TrimPrefix(upstreamAddr, "http://")
+		createdTarget := utils.DefaultTarget()
+		createdTarget.Target = strings.TrimPrefix(utils.UpstreamAddr, "http://")
 		_, err = kongCli.Targets().CreateFromUpstreamId(kongUpstream.Id, createdTarget)
 		gomega.Expect(err).To(gomega.BeNil())
 
-		createdTarget.Target = strings.TrimPrefix(upstreamAddr2, "http://")
+		createdTarget.Target = strings.TrimPrefix(utils.UpstreamAddr2, "http://")
 		_, err = kongCli.Targets().CreateFromUpstreamId(kongUpstream.Id, createdTarget)
 		gomega.Expect(err).To(gomega.BeNil())
 
-		createdService := defaultService()
+		createdService := utils.DefaultService()
 		createdService.Host = gokong.String(kongUpstream.Name)
 		kongService, err := kongCli.Services().Create(createdService)
 		gomega.Expect(err).To(gomega.BeNil())
 
-		createdRoute := defaultRoute()
+		createdRoute := utils.DefaultRoute()
 		createdRoute.Paths = gokong.StringSlice([]string{"/get"})
 		createdRoute.Service = gokong.ToId(*kongService.Id)
 		_, err = kongCli.Routes().Create(createdRoute)
 		gomega.Expect(err).To(gomega.BeNil())
 
-		err = testMigrate()
+		err = utils.TestMigrate()
 		gomega.Expect(err).To(gomega.BeNil())
 
 		apisixRespMap := make(map[string]int)
 		kongRespMap := make(map[string]int)
-		cc := &CompareCase{Path: "/get/get"}
+		cc := &utils.CompareCase{Path: "/get/get"}
 		for range [10]int{} {
-			apisixResp, kongResp, err := getResp(cc)
+			apisixResp, kongResp, err := utils.GetResp(cc)
 			gomega.Expect(err).To(gomega.BeNil())
 			apisixRespMap[string(apisixResp)]++
 			kongRespMap[string(kongResp)]++
